Refactor GUI and add configurable tokenizer
