+++ promptpack.spec
@@ -1,11 +1,12 @@
 # -*- mode: python ; coding: utf-8 -*-
 
+block_cipher = None
 
 a = Analysis(
-    ['promptpack.py'],
-    pathex=[],
+    ['promptpack/__main__.py'],
+    pathex=['.'],
     binaries=[],
-    datas=[],
+    datas=[('promptpack/locales/*.json', 'promptpack/locales')],
     hiddenimports=[],
     hookspath=[],
     hooksconfig={},
