+++ promptpack/tokenizer.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from .settings import load_settings
+
 try:
     import tiktoken  # type: ignore
 except Exception:  # pragma: no cover - optional dependency
@@ -44,17 +46,28 @@ def gemini_tokens(text: str) -> int:
     return info.get("token_count", len(info.get("tokens", [])))
 
 
+
 MODEL_DISPATCH = {
     "gpt": gpt_tokens,
     "claude": claude_tokens,
     "gemini": gemini_tokens,
 }
 
+_DEFAULT_MODEL = load_settings().get("token_model", "gpt")
+_OVERRIDE_MODEL: str | None = None
 
-def estimate_token_count(text: str, model: str = "gpt") -> int:
+
+def set_default_model(model: str | None) -> None:
+    """Override default model used by :func:`estimate_token_count`."""
+    global _OVERRIDE_MODEL
+    _OVERRIDE_MODEL = model
+
+
+def estimate_token_count(text: str, model: str | None = None) -> int:
     """Return number of tokens for the chosen model."""
-    func = MODEL_DISPATCH.get(model)
+    chosen_model = model or _OVERRIDE_MODEL or _DEFAULT_MODEL
+    func = MODEL_DISPATCH.get(chosen_model)
     if func is None:
-        raise ValueError(f"Modello sconosciuto: {model}")
+        raise ValueError(f"Modello sconosciuto: {chosen_model}")
     return func(text)
